utils/calendar_sync.py
"""
Google Calendar synchronization functionality.
"""

from typing import List, Dict
from .logger import LoggerMixin

# Optional Google Calendar integration
try:
    from google.oauth2.credentials import Credentials
    from google_auth_oauthlib.flow import InstalledAppFlow
    from google.auth.transport.requests import Request
    from googleapiclient.discovery import build
    GOOGLE_CALENDAR_AVAILABLE = True
except ImportError:
    GOOGLE_CALENDAR_AVAILABLE = False


class CalendarSync(LoggerMixin):
    """Google Calendar synchronization functionality."""
    
    def __init__(self):
        self.service = None
        self.calendar_id = 'primary'  # Use primary calendar
        
        if GOOGLE_CALENDAR_AVAILABLE:
            self.setup_google_calendar()
        else:
            self.logger.warning("Google Calendar integration not available. Install google-api-python-client for calendar sync.")
    
    def setup_google_calendar(self):
        """Setup Google Calendar API credentials."""
        # TODO: Implement OAuth2 flow
        self.logger.info("Google Calendar sync not yet configured")
        pass
    
    def sync_events(self, events: List[Dict]) -> int:
        """
        Sync events to Google Calendar.
        
        Args:
            events: List of event dictionaries to sync
        
        Returns:
            Number of events successfully synced
        """
        if not GOOGLE_CALENDAR_AVAILABLE or not self.service:
            self.logger.warning("Google Calendar sync not available")
            return 0
        
        # TODO: Implement calendar event creation
        self.logger.info(f"Would sync {len(events)} events to Google Calendar")
        return 0
    
    def is_available(self) -> bool:
        """Check if Google Calendar integration is available."""
        return GOOGLE_CALENDAR_AVAILABLE and self.service is not None
