requirements.txt
# Core dependencies
requests>=2.31.0
APScheduler>=3.10.0

# Date and time handling
python-dateutil>=2.8.2

# Optional Google Calendar integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Additional API libraries for sports data
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.11.0

# HTTP client enhancements
httpx>=0.24.0  # Alternative to requests with async support
aiohttp>=3.8.0  # For async HTTP requests

# Data validation and parsing
pydantic>=2.0.0  # For data validation
jsonschema>=4.19.0  # For JSON schema validation

# Monitoring and metrics
prometheus-client>=0.17.0  # For metrics collection

# Configuration management
python-dotenv>=1.0.0  # For environment variable management

# Logging enhancements
structlog>=23.1.0  # Structured logging

# Time zone handling
pytz>=2023.3

# API rate limiting
ratelimit>=2.2.1

# FastAPI and web dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# CORS support
fastapi-cors>=0.0.6